refactor: symbol을 도메인에서 만들도록 변경
