+++ src/main/java/bowling/BowlingMain.java
@@ -3,9 +3,8 @@ package bowling;
 import bowling.domain.BowlingGame;
 import bowling.domain.Pins;
 import bowling.domain.User;
-import bowling.domain.dto.ResultDTO;
 import bowling.view.InputView;
-import bowling.view.output.OutputView;
+import bowling.view.OutputView;
 
 public class BowlingMain {
     public static void main(final String[] args) {
@@ -13,14 +12,13 @@ public class BowlingMain {
         final User user = User.of(userName);
         
         final BowlingGame bowlingGame = BowlingGame.of(user);
-        final ResultDTO emptyResultDTO = bowlingGame.getResultDTO();
-        OutputView.printResult(emptyResultDTO);
+        OutputView.printResult(bowlingGame.name(), bowlingGame.symbols());
         
         while (bowlingGame.isNotFinished()) {
             final int nowFrameNo = bowlingGame.nowFrameNo();
             final int fallenPins = InputView.plzEnterFallenPins(nowFrameNo);
-            final ResultDTO resultDTO = bowlingGame.pitch(Pins.of(fallenPins));
-            OutputView.printResult(resultDTO);
+            bowlingGame.pitch(Pins.of(fallenPins));
+            OutputView.printResult(bowlingGame.name(), bowlingGame.symbols());
         }
     }
 }
