+++ src/main/java/bowling/domain/frame/FrameFactory.java
@@ -1,4 +1,6 @@
-package bowling.domain;
+package bowling.domain.frame;
+
+import bowling.domain.frame.Frame;
 
 public class FrameFactory {
     public static Frame creates() {
